Add support for PATCH

+++ src/Structures/ApiResourceIdentifier.php
@@ -32,6 +32,11 @@ class ApiResourceIdentifier
     protected $relationships;
 
     /**
+     * @var string[] Array to keep track of relationships that are changed.
+     */
+    protected $changedRelationships = [];
+
+    /**
      * ApiResourceIdentifier constructor.
      *
      * @param string  $resourceType The resource type.
@@ -51,7 +56,7 @@ class ApiResourceIdentifier
      *
      * @return string The resource type.
      */
-    public function type()
+    public function type() : string
     {
         return $this->resourceType;
     }
@@ -61,15 +66,13 @@ class ApiResourceIdentifier
      *
      * @return string|null The resource Id.
      */
-    public function id()
+    public function id() : ?string
     {
         return $this->id;
     }
 
     /**
      * Make a GET request to the resource.
-     *
-     * @throws ExonetApiException The exception.
      *
      * @return ApiResource|ApiResourceSet A resource or resource set.
      */
@@ -81,9 +84,28 @@ class ApiResourceIdentifier
     /**
      * Delete this resource from the API.
      */
-    public function delete()
+    public function delete() : void
     {
-        $this->request->delete($this->id());
+        // If there are no changed relationships, perform a 'normal' delete.
+        if (empty($this->changedRelationships)) {
+            $this->request->delete($this->id());
+
+            return;
+        }
+
+        // If there are changed relationships, transform them to JSON and send a DELETE to the relationship endpoint.
+        foreach ($this->changedRelationships as $relationship) {
+            $relationData = $this->relationship($relationship);
+            if (is_array($relationData)) {
+                array_walk($relationData, function (&$relationItem) {
+                    $relationItem = $relationItem->toJson();
+                });
+            } else {
+                $relationData = $relationData->toJson();
+            }
+
+            $this->request->delete($this->id().'/relationships/'.$relationship, ['data' => $relationData]);
+        }
     }
 
     /**
@@ -101,17 +123,45 @@ class ApiResourceIdentifier
     /**
      * Get a relationship definition to another resource.
      *
-     * @param string $name The name of the relationship.
+     * @param string                                        $name     The name of the relationship.
+     * @param ApiResourceIdentifier|ApiResourceIdentifier[] $resource
      *
-     * @return Relation|Relationship
+     * @return Relation|Relationship|$this The requested relation data or the current resource when setting a relation.
      */
-    public function relationship(string $name)
+    public function relationship(string $name, $resource = null)
     {
-        // Check if the relationship is already defined. If not, create it now.
-        if (!isset($this->relationships[$name])) {
-            $this->relationships[$name] = new Relationship($name, $this->type(), $this->id());
+        // If there are is only a single argument, get the relation.
+        if (func_num_args() === 1) {
+            // Check if the relationship is already defined. If not, create it now.
+            if (!isset($this->relationships[$name])) {
+                $this->relationships[$name] = new Relationship($name, $this->type(), $this->id());
+            }
+
+            return $this->relationships[$name];
         }
 
-        return $this->relationships[$name];
+        // Set the relation data.
+        if (is_array($resource)) {
+            foreach ($resource as $resourceIdentifier) {
+                $this->relationships[$name][] = $resourceIdentifier;
+            }
+        } else {
+            $this->relationships[$name] = $resource;
+        }
+
+        $this->changedRelationships[] = $name;
+
+        return $this;
+    }
+
+    /**
+     * Transform the set identifiers for this resource to an array that can be used for JSON.
+     */
+    protected function toJson() : array
+    {
+        return [
+            'type' => $this->type(),
+            'id' => $this->id(),
+        ];
     }
 }
